refactor: move translation to the i18n module
