+++ api/src/seeder.ts
@@ -13,8 +13,8 @@ import { CategorySeeder } from './chat/seeds/category.seed';
 import { categoryModels } from './chat/seeds/category.seed-model';
 import { ContextVarSeeder } from './chat/seeds/context-var.seed';
 import { contextVarModels } from './chat/seeds/context-var.seed-model';
-import { TranslationSeeder } from './chat/seeds/translation.seed';
-import { translationModels } from './chat/seeds/translation.seed-model';
+import { TranslationSeeder } from './i18n/seeds/translation.seed';
+import { translationModels } from './i18n/seeds/translation.seed-model';
 import { LoggerService } from './logger/logger.service';
 import { NlpEntitySeeder } from './nlp/seeds/nlp-entity.seed';
 import { nlpEntityModels } from './nlp/seeds/nlp-entity.seed-model';
